requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
click==8.1.8
dbus-next==0.2.3
distro==1.9.0
fastapi==0.115.12
groq==0.22.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
mss==10.0.0
numpy==2.2.4
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
pycairo==1.27.0
pydantic==2.11.3
pydantic-core==2.33.1
pygobject==3.52.3
pytesseract==0.3.13
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.1.0
sniffio==1.3.1
starlette==0.46.1
typing-extensions==4.13.1
typing-inspection==0.4.0
uvicorn==0.34.0
websockets==15.0.1
